reduce and consolodate tests
