+++ tests/timer/edge-cases.test.ts
@@ -79,37 +79,7 @@ describe("PomodoroTimer - Edge Cases", () => {
 
 			timer.pauseTimer();
 		});
-
-		it("should format negative time with consistent -MM:SS format", () => {
-			const now = Date.now();
-			const timer = (plugin as PluginWithPrivates)._timer;
-
-			const textEl: any = { textContent: "" };
-			mockStatusBarItem.querySelector = jest
-				.fn()
-				.mockReturnValue(textEl);
-
-			timer.toggleTimer();
-
-			// Test various negative times to ensure consistent formatting
-			const testCases = [
-				{ offset: 1500000 + 1000, expected: "-0:01" }, // 1 second over
-				{ offset: 1500000 + 5000, expected: "-0:05" }, // 5 seconds over
-				{ offset: 1500000 + 60000, expected: "-1:00" }, // 1 minute over
-				{ offset: 1500000 + 65000, expected: "-1:05" }, // 1:05 over
-				{ offset: 1500000 + 600000, expected: "-10:00" }, // 10 minutes over
-				{ offset: 1500000 + 3599000, expected: "-59:59" }, // 59:59 over
-			];
-
-			testCases.forEach(({ offset, expected }) => {
-				jest.setSystemTime(now + offset);
-				(timer as any).updateDisplay();
-				expect(textEl.textContent).toBe(expected);
-			});
-
-			timer.pauseTimer();
-		});
-	});
+});
 
 	it("should handle missing text element in updateDisplay", () => {
 		const timer = (plugin as PluginWithPrivates)._timer;
@@ -136,13 +106,12 @@ describe("PomodoroTimer - Edge Cases", () => {
 		);
 	});
 
-	it("should handle cleanup with multiple registered intervals", () => {
+	it("should handle cleanup with registered intervals", () => {
 		const timer = (plugin as PluginWithPrivates)._timer;
 		const clearIntervalSpy = jest.spyOn(window, "clearInterval");
 
-		// Simulate multiple intervals being registered
+		// Test with multiple intervals
 		timer["registeredIntervals"] = new Set([1, 2, 3]);
-
 		timer.cleanup();
 
 		expect(clearIntervalSpy).toHaveBeenCalledWith(1);
@@ -150,19 +119,10 @@ describe("PomodoroTimer - Edge Cases", () => {
 		expect(clearIntervalSpy).toHaveBeenCalledWith(3);
 		expect(timer["registeredIntervals"].size).toBe(0);
 
-		clearIntervalSpy.mockRestore();
-	});
-
-	it("should handle cleanup with no registered intervals", () => {
-		const timer = (plugin as PluginWithPrivates)._timer;
-		const clearIntervalSpy = jest.spyOn(window, "clearInterval");
-
-		// Ensure no intervals are registered
+		// Test with no intervals
 		timer["registeredIntervals"] = new Set();
-
-		// Should not throw and should not call clearInterval; internal state should remain cleared
 		expect(() => timer.cleanup()).not.toThrow();
-		expect(clearIntervalSpy).not.toHaveBeenCalled();
+		expect(clearIntervalSpy).toHaveBeenCalledTimes(3); // No additional calls
 		expect(timer["registeredIntervals"].size).toBe(0);
 		expect(timer["currentInterval"]).toBeNull();
